checkout routes refactored
